preparing for deploy

+++ standart/main/models/product.py
@@ -7,8 +7,7 @@ class Product(models.Model):
     title = models.CharField(max_length=255, verbose_name='Название товара')
     image_small = models.ImageField(verbose_name='Маленькое фото товара', upload_to='products/', null=True, blank=True)
     image_big = models.ImageField(verbose_name='Большое фото товара', upload_to='products/', null=True, blank=True)
-    price = models.DecimalField(
-        max_digits=7, decimal_places=2,
+    price = models.PositiveIntegerField(
         verbose_name='Цена за 1 комплект',
         null=True, blank=True
     )
